Create Tick Controller Prototype
This Object will locate any conveyor belts and machines inside it's hitbox, and add them to a list.
It will then count down a timer and each time the timer hits zero it will send an activation ping to all objects in list. 
This is so that the machines move on the same interval and the factory looks smoother.